Initial background calculations implementation
Locking does not seem to be working. We need to synchronize on
the non final field root in MCTS instead of putting locks. We
can synchronize on the non final field by introducing a final
field and synchronizing on it instead of root. Currently as it
stands, it does not introduce errors practically but there
could be some edge cases.